requirements.txt
streamlit
transformers==4.46.2
diffusers
accelerate
numpy==1.24.4
audio-recorder-streamlit
ffmpeg-python
edge-tts
langchain
langchain_ollama
beautifulsoup4
Pillow
aiohttp
duckduckgo-search
mysql-connector-python
python-dotenv
plotly
pandas
langchain_google_community